chatbot/requirements.txt
python-dotenv
streamlit
langchain_core
langsmith
langchain_google_genai